+++ frontend/src/store/ApplicationStore.ts
@@ -3,10 +3,17 @@ import { Board } from '../interfaces/Board';
 import { BrowserState } from '../interfaces/BrowserState';
 import { Card } from '../interfaces/Card';
 import { Lane } from '../interfaces/Lane';
+import { ListView } from '../interfaces/ListView';
 import { Schema } from '../interfaces/Schema';
 import { CurrencyCode } from '../interfaces/Team';
 import { User } from '../interfaces/User';
 import { FilterMode } from '../pages/HomePage';
+
+export type InterfaceState =
+  | 'default'
+  | 'card-detail'
+  | 'lane-detail'
+  | 'account-detail';
 
 export interface ApplicationStore {
   users: User[];
@@ -29,7 +36,7 @@ export interface ApplicationStore {
     isPageLoaded: boolean;
   };
   ui: {
-    state: 'default' | 'card-detail' | 'lane-detail' | 'account-detail';
+    state: InterfaceState;
     id: undefined | Card['id'] | Lane['id'] | Account['id'];
     modal: 'success' | 'error' | undefined;
     text: string | undefined;
@@ -37,5 +44,6 @@ export interface ApplicationStore {
       text: string | undefined;
       mode: FilterMode[];
     };
+    accounts: ListView;
   };
 }